extract grade validation from assessment completion into utils function
